MSK Connect
On MSK Connect org.apache.kafka.common.utils.Utils.isBlank is raising an exception since it cannot be found. This might result of an older Kafka used on MSK or the class being modified internally by the AWS team for their distribution.

The code change brings the method inside the lib to avoid the issue found.