Enhancement: Support Record Timestamp for Data Partitioning
This update addresses a common user request to utilize the record timestamp for data partitioning purposes.

Previously, the timestamp converter exclusively worked with the record Key or Value. Users had to resort to additional SMTs to transfer the record timestamp into the payload (Key or Value), or handle it during data production, leading to redundancy.

To streamline this process, this PR introduces support for utilizing the record timestamp directly. It introduces a new `_timestamp` prefix. When specified, the converter will read the record timestamp as the source value for conversion, eliminating the need for additional data manipulation steps.