Timezone support
There have been requests from users to support date/time partition in the S3/GCP/Azure sink with a specific timezone. The code changes brings the timezone configuration as an optional setting. When epoch configuration is involved, changing the timezone to non-UTC yields an error.

Removes the check plugin since the jar dependency is built with a newer JVM version and fails the build.