Connect Runtime fails initialising the static ctor for InsertRollingFieldTimestampHeaders
It's not clear why but the code `RecordFieldTimestamp.extendConfigDef(InsertRollingTimestampHeaders.CONFIG_DEF);` gets Connect plugin scanner throwing an exception. For reasons beyond the code implementation it thinks the code redefines the setting entry named: `field`.

The solution is to redefine the ConfigDef and avoid referencing the static values defined in th parent classes.